Refactored the File Structure. Added Tests for components.
