+++ src/components/AddItem/AddItem.test.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import AddItem from './AddItem';
+import {BrowserRouter} from "react-router-dom";
+
+it('renders without crashing', () => {
+  const div = document.createElement('div');
+  ReactDOM.render(<BrowserRouter><AddItem /></BrowserRouter>, div);
+  ReactDOM.unmountComponentAtNode(div);
+});
