+++ src/components/List/List.js
@@ -1,0 +1,11 @@
+import Navigation from "../Navigation/Navigation";
+
+const List = () => {
+  return (
+    <>
+      <Navigation />
+    </>
+  )
+}
+
+export default List;
